requirements.txt
pycryptodome==3.19.0
requests==2.27.1
